+++ Examples/Get-Greeting.ps1
@@ -1,0 +1,45 @@
+Import-Module '.\AnyBox.psd1'
+
+function Get-Greeting
+{
+    param(
+        [ValidateSet('Hello', 'Get it together', 'Buzz off')]
+        [string]$Greeting = 'Hello',
+        [ValidateNotNullOrEmpty()]
+        [string]$Subject = 'World',
+        [switch]$Exclaim
+    )
+
+    [string]$msg = $Greeting + ' ' + $Subject
+    
+    if ($Exclaim)
+    {
+        $msg = $msg.ToUpper() + '!' 
+    }
+    else
+    {
+        $msg += '.'
+    }
+
+    return $msg
+}
+
+
+[string]$func_name = 'Get-Greeting'
+
+[hashtable]$params = @{
+    Title = $func_name
+    Button = 'OK'
+    MinWidth = 300
+}
+
+[hashtable]$resp = Show-AnyBox @params -PromptsFromFunc $func_name
+
+if ($resp['OK'])
+{
+    [hashtable]$func_params = Get-DictSubset -Source $resp -KeyStartsWith 'param_' -Trim
+
+    [string]$msg = & (Get-Command $func_name).ScriptBlock @func_params
+
+    $null = Show-AnyBox @params -Message $msg
+}